chore: move get_transitive_closure function into the Registry module
