--- docs/source/pages/developers_guide/apidoc/macaron.graph.rst
@@ -1,18 +1,0 @@
-macaron.graph package
-=====================
-
-.. automodule:: macaron.graph
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Submodules
-----------
-
-macaron.graph.graph module
---------------------------
-
-.. automodule:: macaron.graph.graph
-   :members:
-   :undoc-members:
-   :show-inheritance:
