--- src/macaron/graph/__init__.py
@@ -1,4 +1,0 @@
-# Copyright (c) 2024 - 2024, Oracle and/or its affiliates. All rights reserved.
-# Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl/.
-
-"""This package contains classes and functions for generic graph operations."""
